Create User model to store user details
In order to add roles to RODA, we need to first add a User model to
associate these roles with.

For our initial pass, the User instance does not exist until they first
log in. Eventually, the process of creating a User in RODA will also
create an account on Auth0.

To show that this step is working, I'm displaying the user's name on the
dashboard page they see immediately after signing in.