Add IATI type display names as i18n strings
Add the IATI type display names as internationalised & versioned strings. This
means we will be able to update them as IATI is updated to v3 and beyond.

In `application.rb` we are setting the default language as `iati` with a fallback
to `en`. If we need to add specific translations for `beis` we can play with the
translation fallbacks to make the translations cascade as appropriate.