Add 'i18n-tasks' gem
Since we are using i18n strings across the application, it's useful to
have an early warning when we've referenced a string but haven't
defined it in `config/locales/*.yml`.

This change adds an I18n-specific tests which fails if translations are
missing. Running `i18n-tasks missing` shows where the missing key is
used, to ease adding it back.

The alternative is enforcing Rails' 'raise exception on missing
translation', which risks the error being missed in the tests and
blowing up on the user instead.