Record decision to use Heroku
We switched on Heroku without recording why so we are retroactively documenting this.

Even if the choice of hosting on Heroku is intended to be short term our team doesn't have control of how long this will be the case, we are dependant on other teams within BEIS. We should record why we have done this to preserve the rationale for future teams and future questions.