Lock down controllers using Pundit
Now we have policies in place for controlling which actions can be
carried out by users of particular roles, we need to call `authorize` in
every action in order to check that the user have the correct access.

NB: I've added `after_action :verify_authorized` to
`Staff::BaseController`, so that an exception is raised if we forget
to do this.

Tests have been updated to use a user with the administrator role where
appropriate.