Prevent test suite 'reloading is disabled' error
There's a bug with the way Spring interacts with Rails 6. The original
workaround was to set `cache_classes = false` which causes the test
suite to run a *lot* slower, particularly when using Capybara.

An alternative workaround is to use the classic autoloader rather than
Zeitwerk (default in Rails 6), which allows the faster `cache_classes =
true` whilst preventing the annoying reloading errors.

The issue is documented here:
https://github.com/rails/spring/issues/598