Refactor Pundit::NotAuthorizedError handler
Handle `Pundit::NotAuthorizedError` exceptions in the same way as
`UserNotAuthorised` ones in the `Authorisation` concern.

I've also amended the default Pundit error message to be similar to the
one used for `UserNotAuthorised` exceptions.

NB: I've left the Pundit `after_action` in `Staff::BaseController` so
that it's obvious that it's being called without too much digging.