Add GOVUK Design System Form Builder gem
This replaces Rails' default form builder with one built by DFE Digital
to be fully-compliant with the GOVUK Design System. As we have been
using Simple Form up to now, the two can co-exist, which makes migration
easier.

NB: I18n support is coming in the v1.1 release of this gem, which is
currently in development. I've therefore made a fork of this project,
and I'm referencing the `i18n` WIP branch until v1.1 is complete.

Additionally, Rails is configured to submit forms without using XHR.