Create Staff::ActivityFormsController
As a first step, copy the existing form from Staff::ActivitiesController
over to Staff::ActivityFormsController. For now, this is a multi-step
form with only a single step.

Next steps will be to change the Staff::ActivitiesController create
action to redirect to this new form, and then to start breaking the form
down into multiple steps.

Co-authored-by: Dominic Baggott <dominic.baggott@gmail.com>