Start activity creation specs from fund#show
As part of moving to a multi-step form we're getting rid of the fund#new
action. That exposed these specs going straight to the new route, rather
than starting on the show page and clicking the button.

This should make no difference to the specs, just make them more
resilient to other refactoring.

Co-authored-by: Dominic Baggott <dominic.baggott@gmail.com>