Create an empty activity and redirect to wizard
Previously, the activity form was created in its entirety in
`Staff::ActivitiesController`. Now we're creating an empty activity and
immediately redirecting to the first step of the wizard in
`Staff::ActivityFormsController`.

This also removes the `Staff::ActivitiesController#new`, and so the
"Create activity" button needs to post to the `#create` action instead.

Co-authored-by: Dominic Baggott <dominic.baggott@gmail.com>