Welcome emails are only sent after the user is persisted
Now the `user` object is sent to the mailing service (rather than an ID) there is always an initial error from Sidekiq when the user is created.

The error happens when a user is successfully created but the mailer is called before the user record is persisted to the database, due to the use of the transaction block.

When the mailer recieves the object, it tries to call `id` on it and fails with a serialisation error. This is due to the fact no ID has been created and we are dealing with a object in memory only.

Added a missing happy path test on my way through.