Only synchronise with Auth0 when it makes sense
Don't try to synchronise a user with Auth0 if their name and email
address haven't changed.

NB: This change required `user.save` to be moved after the call to
`UpdateUserInAuth0`, as the latter needs to have access to the
`_changed?` methods on `user` to determine if we are updating attributes
that Auth0 needs to know about.