All fund routing requires an organisation
We had routing that allowed fund urls to be used without the context of their organisation.

We had intended to lock this down previously with the following routing block:

```
resources :organisations do
      resources :funds, only: [:index, :show, :new, :create]
 end
```

This is correct.

However we also had:

```
scope module: "staff" do
  resources :funds, only: [] do
      resources :activities, only: [:new, :create, :show]
      …
```

Which kept the old restful routing for /fund. We'd like to keep funds within the context of the organisation that they are attached to help proide enough state to navigate through the different CRUD actions.