Travis reports to Coveralls rather than GitHub Actions
It doesn't make sense for both Travis and GitHub Action to build and set up a testing environment, which is required to generate a coveralls report. We don't want to overhead of having to maintain to test environment configurations. Given we are using Travis already, let's make Travis work. I needed to set CI to true which is further to the Travis docs.