Simplecov warns locally when coverage is <98%
Previously, we would only know if code coverage had dropped when we
raised a pull request and waited for Coveralls to report. Under the
hood, this uses simplecov, so configure that to warn when running the
test suite locally.

Additionally, the coverage report is saved to `coverage/` in addition to
being sent to Coveralls.