Only generate coverage report when all tests run
There doesn't seem to be an obvious way to have simplecov run only when
the full test suite is called, which means it'll warn you of low test
coverage even when only a subset of test are ran.

This change tries to minimise the chances of this happening, by only
invoking simplecov when there are more than 3 *types* of tests running
(determined by their spec path).