Refactor tracking when the wizard is complete
It appears that the original tracking of a completed wizard (checking
that an activity's `wizard_status` is `complete`) was never fully
implemented, which meant that the current step on the activity form was
often inconsistent.

The idea is that the `wizard_status` tracks the current step when
you initially progress through the form. After you reach the final step,
it will be changed to `complete`, and any further edits will redirect
back to the show page immediately after successing a valid form for the
currently edited step.

This is the first step - introducing a method to check
`wizard_complete?`.