Dates updated to have human readable format as default
We have set  a default date format in translation locale. Added presenters for all the fund activity dates and included tests checking if those specific dates are in a human readable format.

We used ruby travel_to helper method in our tests to ensure that they will continue to pass without being dependent on a current date.  Especially becaue we know that activity dates are validated to be within a certain time frame.

We thought about using strftime in our helper, but then we realised it would’ve been duplicated in a lot of places. Therefore setting the date format in locale file means we don’t need to worry about using locale in the views for an activity and it will be easier to change our time formats in the future.

Co-authored-by: Tom <tomhipkin@gmail.com>