Re-introduce programmes
Having refactored the underlying model, we re-introduce the idea of an
Activity being a 'programme'.

At this point a programme can be created but has no relation to any
other activity, that will come later.

We also expect to change where and how a programme is created and
associated to other activities.

Co-authored-by: Laura Porter <laura@dxw.com>