Make the activity show back link smart
Now we have two levels of activity, where the back link takes a user
will depend on what the 'current' activity is.

This commit introduces the idea of a 'parent' activity, we don't feel
this is the best approach to this idea, but right now, we only need it
to work for a back link so we are happy to go with it.

With this change any programme level activity will need to have a
'parent' fund level activity in order for the show page to render
correctly.