Adjust feature specs and factories
Since the behaviour of the stepped forms has changed the feature specs
have been updated.

Made the `at_identifier_step` trait reflect the behaviour by setting
identifier to nil and moved the `at_purpose_step` trait to below the
`at_identifier_step` which feels like a better place for it.

Removed one feature spec that no longer reflects the behaviour and added
coverage for the two states of the activity form:

- complete where we go back to the summary after editing
- incomplete where we go to the next step in the stepped form