Expose the Accountable Organisation details in the Activity XML
I have updated the XPath for the `participating-org` tests in
`spec/features/staff/fund_managers_can_view_an_activity_as_xml_spec.rb` - instead
of checking the values of the `@role` attribute, I am using it as an indicator
of whether the `participating-org` values are for the _funding_ organisation
(@role = 1) or _accountable_ organisation (@role = 2)

As these @role attribute values are hardcoded in the HAML, there's no point in
checking the values are correct. However we do need to check that the values
in the attributes and children of `participating-org` nodes vary and are correct
according to their @role value.