Fix organisation factory
`default_currency` should be supplied in capital letters to work as
expected.

We also use this factory to seed the database and so we were creating a
invalid organisation as default_currency is a required field.

We also had a validation spec that relied on the fact that the
default_currency was invalid which I found a little unclear, so I have
updated to it to deliberately supply and empty string to a required
field, which I find more explicit.