Add IATI reference to organisations
The IATI reference is used to uniquely identify an organisation and is
published alongside any references to the organisation.

As there is not a formal standard for the reference we can only follow the
IATI recommendations for creating the reference. This also makes the
reference difficult to validate, at this point we collect it from a text
field without validation.

See the IATAI documentation:

https://reference.iatistandard.org/organisation-identifiers/

As the IATI reference is required to publish data to IATI we require it
on organisations.