Add accessible action link helper
For action links such as Edit and Add, we want them to provide more context to the element that will be edited when a link is clicked.

To do that we include a span within a link, that is visually hidden but available to the screen readers. It’s a pattern use in gov.uk design system.

So that when a screen reader stops at a link, it will read more meaningful description such as ‘Edit transaction’ instead of general edit.

Our helper uses ruby link_to method but takes an optional third parameter that adds the span with more context description.

Co-authored-by: Meyric <meyric@dxw.com>