Add hint text to date form fields
During our testing the date form fields turned out to be not clear on what values are expected for the month part. It wasn’t clear that the date input expects a number and not a month name.

Now with a hint text this should be more descriptive.

Next todo’s:
There should be more hint text work around dates added to include guidance around planned and actual dates