Fix flakey user helper
This spec failed a couple of times but mostly passes so I took a look
over it.

As we cannot guarantee the order entities from an `all` query the match
could fail if the order came back differently, which we have no control
over.

We could've used `match_array` if the order didn't matter, but I think
it makes sense that, where this is used, the organisations are listed in
alphabetical order.

I've changed the implementation to use the existing `sorted_by_name`
scope, which gives us the organisations in alphabetical order.

I also tweaked the spec now that the order matters.