Pull request links to IATI validator documentation
Before this change, when clicking the link from a GitHub pull request it broke.

The path was:
https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/pull/doc/xml-validation.md

Notice the /pull/doc/, I guess the relative syntax works from the context of the app but doesn't work from the context of a pull request.

An absolute link is prone to breaking if we change our master branch or move the file. In which case a pointer to documentation that exists is still useful even if the link breaks.