Activity dates have correct roles
When validating an activity that only had planned dates there was an error that stated the activity start date could not be in the future.

I believe this was because we had rendered the planned_end_date with the role intended for the actual_start_date. This meant actual_start_date was being run incorrectly.

IATI activity-date roles defined here http://reference.iatistandard.org/202/activity-standard/iati-activities/iati-activity/activity-date/