Only users that belong to BEIS can manage users of the service
Instead of relying on the role of the user (which we are hoping to factor away) we rely on the type of organisation they are associated with.

Every user must be associated with at least one organisation, either BEIS or a delivery partner. This is marked by the use of the `service_owner` boolean for the moment.

Whilst we may want to break down what users can do within organisations on a read or write basis in the future. For private beta we would like a simpler world where any user of an organisation has full priviledges. If BEIS they can administer users, but DP cannot.

This commit squashes two very similar feature tests into a single file: editing a user