Only users that belong to the BEIS organisation can manage other organisations
Users who only belong to an organisation can still see and edit their own.

This change highlighted several tests that were dependant on the clicking the "manage organisations" button. Since we have refactored the behaviour of the dashboard so that it logs the user into their organisation show page, this no longer needs to happen and is a redundant step.

Rename the organisation feature tests where only BEIS users should ever be able to deal interact with it for clarity.