Only BEIS users can see and interact with funds
Use a custom FundPolicy to help us show and hide buttons to users rather than the generic ActivityPolicy. This allows us to keep the logic of what level an activity is out of the ActivityPolicy. We ended up going down this route and found the code hard to reason with.

We may end up refactoring the ActivityController into level based controllers at a later date as this may further help clean up the views. For example, FundController would do #show and #index as well as #create.

Continued the work to refactor the test suite to rely on the users organisation rather than the role. Deferring to the factory to know how to create the right type of organisation when creating a :beis_user or a :delivery_partner_user. This clears up the tests by removing many organisation initialisaters which are all trying to remember how to create organisations in the right way.