A user can report project level activities
This work is purposefully the simplest implementation of this user need.

Projects are children of programmes. When viewing a programme a user can
add project level activities. The project level activity has the
following attributes:

- organisation (reporting organisation) is BEIS
- funding organisation is BEIS
- accountable organisation is BEIS
- extending organisation is the current user's delivery partner
  organisation

Later work will build on this to allow delivery partners  to be the
accountable organisation for non government delivery partners and
authorise what users can see and do on activities.

Includes a small refactor of the `is_<some level>_level?` helpers. As
the enum provides helper methods we do not have to declare them. We can
also user a prefix and suffix to keep the nice, explicit intention.
(thanks @lozette).