Only BEIS users can manage transactions on a programme
During a large and hairy rebase on top of develop this commit flagged
big differences between approaches. It forced me to tackle the
way we were rendering content on activity#show using enum methods such
as fund? and programme? as well as who sees what.

Since it was very hard to understand who was seeing what content in
order to fix the tests, this test also includs a refactoring that moves
content from the activity#show into reusable partials.