BEIS users can create budgets on a fund and programme, delivery partner users cannot
The budget policy has been tidied up by removing the conditionals and replacing them with a simple 'true'. This is due to the fact that now all users are administrators so there's no longer a meaningful check here.

I'm not sure what the `user.administrator? || activity?` check was meant to be doing. It effectively says ignore the user role and if an activity association exists - return true. Since to have transactions an activity will always exist I'm not sure what this was doing. Removing it didn't break any tests which gave me confidence.