Authorise who can see and interact with activities
Before this change there was nothing stopping users changing the URL and being able to jump to activities they shouldn't be able to. This was because everyone was an administrator! Now we've hidden the buttons in the view, we also need to protect the end points from being changed and restrict access.

Instead of using lots of feature tests for these unhappy paths I'm using lots of faster policy unit tests.