Add implementing organisations to the Activity XML
Expose the implenting organisations as `participating-org[@role = '4']` in the
XML.

This was proving tricky to test as we do not know the order the implwnwting organisations
will be added to the activity record and the XML. So we have decided to test to
see if the implementing organisation data is present, and not worry about testing
the order of the `participating-org` elements. This order does not matter to the
IATI validator anyway.