Only delivery partners can create projects, BEIS users cannot
In doing this work I corrected a bug where projects didn't belong to the delivery partner, rather they belonged to BEIS!  Our create activity buttons always set the organisation to that of the parent activity. Whilst this happened to work for funds and programmes (since they are created by the same organisation, it doesn't work for projects as they shouldn't belong to the same organisation that created the programme.