Fix Docker start script so that containers start OK when the database is new
With the previous version of the code when a new container started rake was being instructed to create a new database, assuming one did not exist. For real environments such as Heroku and GPaaS, the provisioning of a database resource creates the database for us and then provides the DATABASE_URL. This causes the container to crash as it fails to create a database where one already exists.

I presume this was once used when containers were used locally for convenience. For this app we are not running with docker in development.

We recently found this problem when resetting our Heroku database and we've now noticed it again when setting up a new environment on GPaaS.

Given we know databases are always provisioned outside of the Docker context (either by hand or by infrastructure as code) then we do not need the docker entrypoint to handle this responsibility.

Rails 6 introduces a new command that handles the problem this code was trying to fix. The fact that if no database exists the database command errors. Here's a blog post that explains the situation and what rake db:prepare brings: https://blog.saeloun.com/2019/04/11/rails-6-rails-db-prepare.html

I think we can use this command to create the database, run the seed file (we have guarded this from being run outside of development) and then any migrations safely.