Extract env vars from VCAP_SERVICES
GPaaS provides a VCAP_SERVICES environment variable which contains JSON
with information about services bound to the application.

From the JSON we need to extract the REDIS_URL and any other environment
variables we have put into a user provided service.

This is based on the work we had done for CCS RMI
https://github.com/dxw/DataSubmissionServiceAPI/pull/376 and others