Activity and Organisation factory can always be created without conflicting on identifiers
When using factories locally I ran into problems creating new records because the identifier had already been used.

For me a good factory can be created without having to worry about the state of the rest of the database. Although this is mostly handled in our test suite by using DatabaseCleaner to wipe it between tests, I think this quality should extend to using it on the Rails console, or in our seed files.