Activities share their IATI identifier that includes the reporting organisation idenifier
The IATI specification says that the `iati-identifier` should include the reference of the `reporting-org`: http://reference.iatistandard.org/202/activity-standard/iati-activities/iati-activity/iati-identifier/. This change puts this value into the XML format for an activity.

The approach I've taken is for an activity to keep a copy of the reporting organisations reference. I was going to keep it dynamic and use `activity.organisation.iati_reference` to populate this field however I read that there activities MUST not change their identifier once published. Given that government organisations can be merged or created these identifiers may in future change, and this service currently allows IATI references to be changed very easily.

I thought about prefixing the existing `activity.identifier` field with the value of the organisation during the creation but this has problems too. If we prefix it, then display it back to the user in the #show views our users may well be confused by the magical prefix and try to correct the information they see in front of them.

In future when we approach publishing we may want to set this value at the point of publish rather than the point of creation.

A new XML presenter is added to allow view customisations that are only relevant for XML views.