An activity can tell us all of it's parent activities
I needed this functionality to help create an accurate IATI identifier that included the parents when present but found it missing.

There are probably smarter ways to query up a tree, perhaps with recursion but I found I was getting bogged down in implementation detail. This works, and we know we will have a finite number of levels to our tree (max 4 with the addition of sub-projects).

The choice of name to associate an activity to it's parent being `activity.activity` is again showing that it might make more sense to the reader if it were `activity.parent` or ` activity.parent_activity`. That said, this is not the focus of this work so I am working within the old naming convention.