iati-identifier for programme and projects include their parent activities
The IATI guidance on the `iati-identifier` field does not explicitly mention combining activity ids from activities in an internal hierarchy: http://reference.iatistandard.org/202/activity-standard/iati-activities/iati-activity/iati-identifier/ - so this is not straight forward.

At the moment BEIS publish records to IATI with iati-identifiers that look something like this: "GB-GOV-13-GCRF-UKSA_NS_UKSA-019" which you can find in this list of real data http://preview.iatistandard.org/index.php?url=https%3A//science-and-innovation-network.s3.eu-west-2.amazonaws.com/UKSA%2BR%2526I%2BActivity%2BFile.xml.

The identifier of the activity is constructed with more than what the delivery partner provides, it is composited with how else BEIS identifies the activity.

GB-GOV-13 = BEIS as the reporting org
GCRF = Fund activity identifier
UKSA_NS_UKSA-019 = a composite programme/project level ID that is currently not confidently understood or mapped to in the trackers.

The Fund level looks straight forward, we can add that to the identifier.

The Programme and project level will be better defined now we are asking programmes to have their own identifier this _might_ meaningfully translate to `UKSA_NS`.

The project would be 019.