Refactor activity validation methods to become private
I noticed that we had no tests for these methods when adding a new one in.

It looks like they are all only used by this class and I can't imagine why we'd ever need to open this logic up in an external interface to other classes. I think we can make these methods private and allow the unit and valdiation specs to test the logic.