Budget dates are validated
The IATI standard requires that budget start and end dates be validated
as follows:

- start date must not be after end date
- end date not more than one year after start date

IATI does not specify what one year means, so we decided to make it 365
days exactly.

By using 365 days we can also avoid the problem of a leap year. i.e.
when a year is 366 days long. We've been clear in the UI that the date
should be no more than 365 days apart and we feel that users will equate
365 days = one year, we are happy for user research to test this.

http://reference.iatistandard.org/203/activity-standard/iati-activities/iati-activity/budget/

We spent some time attempting to make a generic validator for these two
validations but in the end decided that a validator specifically for
budget was the simplest and most elegant solution until we need to reuse
this validation criteria.

We decided that as the validation requirements are clearly defined by
IATI, by wrapping them up in a validator we keep all of the elements to
carry out the validation together in one place.

Co-authored-by: MariluzHerrera <maria@dxw.com>