CI will do a final security check using Brakeman in the 'test' step
Brakeman advise not calling using the gem through Rake as it loads the entire application: https://brakemanscanner.org/docs/rake/ I found it to work but it span out a lot of warnings about duplicate loading.

The choice to have this check only on CI was due to the fact that these issues should occur rarely. Having brakeman run on every local test run would introduce a lot of overhead. The main thing is that we catch security vulnerbilities before they go to a real environment. Which I think this should do.

By default Brakeman will error on warnings, errors and it'll be quiet. It will also not print the results to a file when one is not provided.