Reduce the attack footprint by removing unused rails ties
ActiveStorage in default configuration created valid routes that allowed files to be uploaded. We do not intend for files to be uploaded to the application as local files so this can be removed.

If we do want to add file upload in the future we may add this back however it will very likely use a third party document store given the ephemeral nature of Docker containers.

This change also removes action_mailbox routes with allows the application to recieve incoming emails. https://guides.rubyonrails.org/action_mailbox_basics.html This is not appropriate now and I cannot imagine when this would be appropriate.