WIP
Redis version 4 is being used in spite of the GPaaS need for it to be rolled back to Redis 3. There are no supported Redis Docker images for version 3 available for use. We will have to be out of parity here for the moment however we hope GPaaS will have version 5 support within the week.