Temp fix for not working fieldset legends
This is a temporary fix, because there are issues trying to render an h1 legend for a filedset.

Because we removed page title heading from the form wrapper, we need to output it on the fieldset pages. It’s inside the form which is not ideal, but we’re workinng on fixing this.