Add recipent region to the translation
We have switched to use labels as headings in most of the form steps for creating activity, but we are still using page title in our tests.

TODO:
Having two seperate translation files for the app and form, will become make things harder once we start changing a lot of content [ I guess], so we should probably update some of the tests to into the form transaltion for the words we’re loking for.