Hint text changed to be dynamic
When creating or editing an activity, on the status step now we don't have a link to
IATI documentation as we do not need it anymore, but instead we have a help text that
indicates the user the stage their fund,programme or project is (activity.level).
This is now dynamic and it will change to show the right level of activity every time.

This change was made as a requirement from the feedback users gave recently.

I have added a spec to each of the activity levels to make sure the code is working
correctly. In addition, I have deleted the test that checked whether the IATI link was
present as it is not needed anymore.