Remove TRAVIS_BRANCH from docker tag
You can not have a '/' in a docker tag because it uses it to separate org
and repo

```
invalid argument "thedxw/beis-report-official-development-assistance:feature/287-create-ability-to-make-a-user-eg-delivery-partner-inactive-in-the-application" for "-t, --tag" flag: invalid reference format
```

We often use '/' in branch names which causes this error.

We are using this tag to make sure we have a unique tag for deployments
so TRAVIS_BUILD_NUMBER and TRAVIS_COMMIT will suffice.