Add geography step
The IATI standard defines 'geography' as the term used to describe where
in the world will benefit from an activity.

http://reference.iatistandard.org/201/activity-standard/overview/geography/

An activities geography can be one of these three:

- recipient_region
- recipient_country
- location

At the moment we only capture recipient_region and recipient_country.

This commit adds a step to capture which geography will benefit so that
future work can branch the stepped form based on the users choice
capturing either recipient_region OR recipient_country.

Whilst 'geography' may not be the language to describe this data, it is
documented in the IATI standard and so I feel it is a good choice for
the code, even if the content shown to the user changes at a later
point.