Branch stepped form based on geography
When a users chooses recipient_region geography, they skip the
recipient_country step.

When a users chooses recipient_country geography, they skip the
recipient_region step.

A user can change the geography and then choose the correct region or
country. We know this may not be the best experience, but it does allow
the user to complete the task at hand and keeps the changes in the code
small and focused.

Validating the model with the branching in place has proved to be
difficult. This work removes `wizard_complete` from the recipient_region
and recipient_country conditional validations. Whilst this may allow the
attributes to be missing, there is now way in the app to actually
achieve the state required.Until we fully understand when an activity
must be valid (most likely when 'submitted') this change will suffice.

Future work will look at replacing the country and region select boxes
with something else and is likely to resolve this issue.