Do not add any codelist items with status "withdrawn" to the select array
We oticed some duplicate items in the "sector" dropdown list when creating an
activity. It transpired that some of the items in the codelist are marked as
"status: withdrawn" and therefore should not be used in activity creation.

This PR does not put the withdrawn items into any codelist select box.