Stop running the test script when a command errors
Using `set -e` at the top of the script ensures that if one of the
following commands errors the script will stop and return an exit code
of not zero which will let CI know that something has failed.

Otherwise a command will fail during the script but Travis thinks
everything is green.

Also set `u` so if a variable is unbound it will also error out.

At the end of the script do `set +eu`  so that errors are allowed later.
When doing deploys the next command returns an error which is expected.