yml_to_status_objects method created
In order to parse the yaml file and pass it to the view file I created a new method
on the codelist_helper that is specific for status yaml data.
The reason for this is because the hint text on the radio buttons is given by the `:description` field
in the yaml file, which needed to be included when reading this file but does not need to be included
when reading other yaml files (as this creates conflicts and results in tests throwing errors).
In addition, I thought that sorting the different status levels by code
(order in which they will happen) instead of by name (alphabetical order)
would be more appropiate as they now appear in a more "chronological" order, improving the UX in my opinion.

I also added some unit tests for this new method.