Status selection box are now radio buttons
As part of the feedback we recently got from user research, the selection box (dropdown menu)
on the status stage of an activity should be displayed as radio buttons with hints instead.

This change has been implemented in the status view.

Some of the tests needed to be changed as the data is not on a selection box anymore and the way
Capybara navigates through radio buttons is slightly different. I also added assertions
for all the hints displayed on the Status page.