Catch ActiveRecord::MultiparameterAssignmentErrors when date values are zero
I'm not sure this is the best way to catch these errors, so this is a "first
pass" attempt.

When one of the date 3 date values from the UI is "0", Rails attempts to parse
the three values into a date before assigning it to the record, but throws a
ActiveRecord::MultiparameterAssignmentErrors error. Catch this error and use it
to display an error message on the dates page.

This is imperfect because we do not know *which* date the error originates from
so we're not able to show on the UI the exact field which generated the error.